Upgrade iso8601 dependency from 0.1.8 to 0.1.9.
This release fixed an issue in which input strings which started
with a valid, complete iso8601 date (but contained arbitrary other
data after the valid prefix) would be parsed as valid, ignoring
the trailing arbitrary data.

This was causing some id strings that begin with YYYYMM to be
converted into datetime objects (discarding the rest of the id)
instead of remaining as strings. For example,
  '20151102bf08d0ad93eb49be0e33e8c675fee3' => datetime.datetime(2015, 11, 2, ...)